+++ src/Service/CharacterService.php
@@ -1,0 +1,33 @@
+<?php
+namespace App\Service;
+use DateTime; 
+use App\Entity\Character;
+use Doctrine\ORM\EntityManagerInterface;
+
+class CharacterService implements CharacterServiceInterface
+{
+
+    public function __construct(
+        private EntityManagerInterface $em,
+    ) {}
+    // Creates the character
+    public function create(): Character
+    {
+        $character = new Character();
+        $character->setKind('Dame');
+        $character->setName('Anardil');
+        $character->setSlug('anardil');
+        $character->setSurname('Amie du soleil');
+        $character->setCaste('Magicien');
+        $character->setKnowledge('Sciences');
+        $character->setIntelligence(180);
+        $character->setStrength(180);
+        $character->setImage('/dames/anardil.webp');
+        $character->setCreation(new DateTime());
+
+        $this->em->persist($character);
+        $this->em->flush();
+
+        return $character;
+    }
+}