+++ src/Service/CharacterServiceInterface.php
@@ -1,0 +1,8 @@
+<?php
+namespace App\Service;
+use App\Entity\Character;
+interface CharacterServiceInterface
+{
+    // Creates the character
+    public function create(): Character;
+}